move code formatting into components, remove tests file, add interactive propr
