+++ rollup.config.js
@@ -6,7 +6,6 @@ import outputManifest from 'rollup-plugin-output-manifest';
 const componentFiles = [
 	'./pages/assets/base-imports.js',
 	'./pages/assets/code-demo-wrapper.js',
-	'./pages/assets/interactive-demo-wrapper.js',
 	'./pages/assets/navigation.js',
 	'./pages/assets/resizable-demo.js'
 ].concat(files);
